Device Inventory V2: File renamed.
