+++ cyences_app_for_splunk/bin/device_inventory_v2_util.py
